refactor(decompression): move animated rotation unpacking function
